+++ test/org/zaproxy/zap/extension/spiderDSStore/parser/DsStoreHeaderUnitTests.java
@@ -20,27 +20,27 @@ public class DsStoreHeaderUnitTests {
     DsStoreHeader mockedDsStoreHeader = mock(DsStoreHeader.class);
 
     // Valid Byte Streams
-    final Byte[] resourceHeaderInitialisation = {0x00,0x00,0x00,0x01};
-    final Byte[] resourceHeaderMagicByteSequence1 = {0x42,0x75,0x64,0x31};
-    final Byte[] resourceHeaderOffset = {0x00, 0x00, 0x02, 0x00};
-    final Byte[] rootBlockSize = {0x00, 0x00, 0x08, 0x00};
-    final Byte[] resourceHeaderUnknownBlocks ={0,1,2,3,4,5,6,7,8,9,0,1,2,3,4,5};
+    final Short[] resourceHeaderInitialisation = {0x00,0x00,0x00,0x01};
+    final Short[] resourceHeaderMagicByteSequence1 = {0x42,0x75,0x64,0x31};
+    final Short[] resourceHeaderOffset = {0x00, 0x00, 0x02, 0x00};
+    final Short[] rootBlockSize = {0x00, 0x00, 0x08, 0x00};
+    final Short[] resourceHeaderUnknownBlocks ={0,1,2,3,4,5,6,7,8,9,0,1,2,3,4,5};
 
     // Full Valid Stream              | Init              | Bud1              | RootBlock Offset  | Root Block size   | Offset            | unknown Blocks                |
-    final Byte[] fullResourceHeader = {0x00,0x00,0x00,0x01,0x42,0x75,0x64,0x31,0x00,0x00,0x02,0x00,0x00,0x00,0x08,0x00,0x00,0x00,0x02,0x00,0,1,2,3,4,5,6,7,8,9,0,1,2,3,4,5};
+    final Short[] fullResourceHeader = {0x00,0x00,0x00,0x01,0x42,0x75,0x64,0x31,0x00,0x00,0x02,0x00,0x00,0x00,0x08,0x00,0x00,0x00,0x02,0x00,0,1,2,3,4,5,6,7,8,9,0,1,2,3,4,5};
     final int offsetSizeOfFullResourceHeader = 0x0200;
     final int rootBlockSizeOfFullResourceHeader = 0x0800;
 
 
     // Offset for InvalidOffsetTest
-    final Byte[] invalidResourceHeaderOffset = {0x00, 0x00, 0x04, 0x00};
+    final Short[] invalidResourceHeaderOffset = {0x00, 0x00, 0x04, 0x00};
 
     Logger logger = Logger.getLogger(DsStoreHeaderUnitTests.class);
 
     @Test
     public void vaidateHeaderWithCorrectHeader_Unittest(){
         // given
-        Byte[] fullvalidHeader=null;
+        Short[] fullvalidHeader=null;
         // when
         try{
             fullvalidHeader = ByteUtils.mergeByteArrays(resourceHeaderInitialisation,resourceHeaderMagicByteSequence1, resourceHeaderOffset, rootBlockSize, resourceHeaderOffset, resourceHeaderUnknownBlocks);
@@ -55,7 +55,7 @@ public class DsStoreHeaderUnitTests {
     @Test
     public void validateHeaderWithInvalidOffset_Unittest(){
         // given
-        Byte[] fullInvalidHeader=null;
+        Short[] fullInvalidHeader=null;
         // when
         try{
             fullInvalidHeader = ByteUtils.mergeByteArrays(resourceHeaderInitialisation,resourceHeaderMagicByteSequence1, null, rootBlockSize, null, resourceHeaderUnknownBlocks);
@@ -70,7 +70,7 @@ public class DsStoreHeaderUnitTests {
     @Test
     public void validateHeaderWithNullRootBlock_Unittest(){
         // given
-        Byte[] fullResourceHeader=this.fullResourceHeader;
+        Short[] fullResourceHeader=this.fullResourceHeader;
         fullResourceHeader[(int) Math.floor(fullResourceHeader.length/2)]=null;
         fullResourceHeader[(int) Math.floor(fullResourceHeader.length/3)]=null;
         // when
